Move dotnet-fpe repo into dotnet repo
Former-commit-id: b8adfb6838f7337398dc61cd2f17659ded263de0

+++ src/UbiqSecurity/Billing/IBillingEventsManager.cs
@@ -5,7 +5,7 @@ namespace UbiqSecurity.Billing
 {
 	internal interface IBillingEventsManager : IDisposable
 	{
-		int EventCount { get; }
+		long EventCount { get; }
 
 		Task AddBillingEventAsync(string apiKey, string datasetName, string datasetGroupName, BillingAction billingAction, DatasetType datasetType, int keyNumber, long count);
 	}
